+++ app/assets/javascripts/application.js
@@ -1,3 +1,2 @@
-
 //= require rails-ujs
 //= require_tree .
