+++ app/javascript/packs/map.js
@@ -13,10 +13,7 @@ if (mapElement) { // don't try to build a map if there's no div#map to inject in
   } else {
     map.fitLatLngBounds(markers);
   }
-  // debugger
-}
-
-const styles = [
+  const styles = [
     {
         "featureType": "water",
         "elementType": "geometry",
@@ -212,3 +209,7 @@ styles: styles,
 mapTypeId: 'map_style'
 });
 map.setStyle('map_style');
+
+  // debugger
+}
+
