+++ app/views/flats/show.html.erb
@@ -84,3 +84,7 @@
   <div id="img-caption"></div>
 </div>
 
+<% content_for :after_js do %>
+  <script> modalPopup(); </script>
+<% end %>
+
